refactor: extract useSignleMatchPageData hook
